chore: major refactor of threshold-fhe crate
