+++ core/service/src/util/key_setup/mod.rs
@@ -945,7 +945,7 @@ async fn ensure_ca_cert_exists<PubS: Storage>(
         &rcgen::PKCS_ECDSA_P256K1_SHA256,
     )?;
     let (ca_cert_ki, ca_cert, _ca_params) =
-        threshold_fhe::tls_certs::create_ca_cert_from_ca_keypair(
+        threshold_networking::tls_certs::create_ca_cert_from_ca_keypair(
             subject.as_str(),
             tls_wildcard,
             &ca_keypair,
