+++ core/threshold-experimental/Cargo.toml
@@ -7,77 +7,25 @@ authors.workspace = true
 license.workspace = true
 description = "Experimental BGV/BFV FHE schemes with distributed key generation, threshold decryption, and supporting algebraic primitives (NTT, cyclotomic rings, CRT)."
 
-[[bin]]
-name = "stairwayctl"
-path = "src/bin/stairwayctl.rs"
-required-features = ["choreographer", "threshold-fhe/choreographer"]
-
-[[bench]]
-name = "bgv"
-harness = false
-# required-features = ["experimental", "testing"]
-
-[[bench]]
-name = "algebra"
-harness = false
-# required-features = ["experimental"]
-
-[[bench]]
-name = "decoding"
-harness = false
-# required-features = ["experimental", "extension_degree_8"]
-
-
-# BGV Benchmarks
-[[bench]]
-name = "non-threshold_bgv_basic-ops_speed"
-path = "benches/non-threshold/bgv/speed/basic_ops.rs"
-# required-features = ["experimental"]
-harness = false
-
-[[bench]]
-name = "non-threshold_bgv_keygen_speed"
-path = "benches/non-threshold/bgv/speed/keygen.rs"
-# required-features = ["experimental"]
-harness = false
-
-[[bench]]
-name = "non-threshold_bgv_basic-ops_memory"
-path = "benches/non-threshold/bgv/memory/basic_ops.rs"
-# required-features = ["measure_memory","experimental"]
-harness = false
-
-[[bench]]
-name = "non-threshold_bgv_keygen_memory"
-path = "benches/non-threshold/bgv/memory/keygen.rs"
-# required-features = ["measure_memory","experimental"]
-harness = false
-
-
 [dependencies]
 aes-prng.workspace = true
 algebra.workspace = true
-async-trait = { workspace = true, optional = true }
 anyhow.workspace = true
 bc2wrap.workspace = true
 clap.workspace = true
 crypto-bigint.workspace = true
-dashmap = { workspace = true, optional = true }
 error-utils.workspace = true
-futures = { workspace = true, optional = true }
 threshold-execution = { workspace = true, features = ["non-wasm"] }
 hashing.workspace = true
 itertools.workspace = true
 threshold-networking.workspace = true
 num-integer.workspace = true
-observability = { workspace = true, optional = true }
 paste.workspace = true
 rand.workspace = true
 serde.workspace = true
 tfhe-csprng.workspace = true
 tfhe.workspace = true
 thread-handles.workspace = true
-threshold-fhe.workspace = true
 threshold-types.workspace = true
 tokio.workspace = true
 tonic.workspace = true
@@ -85,13 +33,6 @@ tracing.workspace = true
 
 [dev-dependencies]
 threshold-execution = { workspace = true, features = ["non-wasm", "testing"] }
-criterion = { workspace = true, features = ["async_tokio", "html_reports"] }
-pprof.workspace = true
-rayon.workspace = true
-algebra = { workspace = true, features = ["extension_degree_8"] }
-threshold-fhe = { workspace = true, features = ["measure_memory", "non-wasm"] }
-peak_alloc = { workspace = true }
 
 [features]
-choreographer = ["threshold-fhe/choreographer", "threshold-fhe/non-wasm", "dep:dashmap", "dep:futures", "dep:async-trait", "dep:observability"]
 slow_tests = []
