+++ core/threshold-experimental/src/bgv/ddec.rs
@@ -58,7 +58,7 @@ fn partial_decrypt<N: Const + NTTConstants<LevelOne>>(
 }
 // run decryption with noise flooding
 #[instrument(name = "BGV.Threshold-Dec", skip_all,fields(sid = ?session.session_id(), my_role = ?session.my_role()))]
-pub(crate) async fn noise_flood_decryption<
+pub async fn noise_flood_decryption<
     N: Clone + Const + NTTConstants<LevelOne>,
     S: SmallSessionHandles<LevelOne>,
 >(
