+++ core/threshold-experimental/src/constants.rs
@@ -14,9 +14,8 @@ pub(crate) const LOG_PLAINTEXT: u32 = 17;
 /// B bound for the NewHope distribution.
 pub(crate) const NEW_HOPE_BOUND: usize = 1;
 
-#[cfg(feature = "choreographer")]
 /// Input party acting as dealer in the (dummy) keygen protocol.
-pub(crate) const INPUT_PARTY_ID: usize = 1;
+pub const INPUT_PARTY_ID: usize = 1;
 
 pub static PLAINTEXT_MODULUS: LazyLock<NonZero<Limb>> =
     LazyLock::new(|| NonZero::new(Limb(65537)).unwrap());
