+++ core/threshold-networking/Cargo.toml
@@ -10,6 +10,8 @@ description = "Provides networking for all threshold network components. GRPC, s
 [dependencies]
 algebra.workspace = true
 anyhow.workspace = true
+clap = { workspace = true, features = ["derive", "env"] }
+rcgen.workspace = true
 async-trait.workspace = true
 attestation-doc-validation.workspace = true
 backoff = { workspace = true, features = ["tokio"] }
@@ -28,6 +30,7 @@ tfhe-versionable.workspace = true
 thread-handles.workspace = true
 threshold-types.workspace = true
 tokio = { workspace = true, features = [
+    "fs",
     "sync",
     "rt",
     "macros",
@@ -42,6 +45,7 @@ tracing.workspace = true
 x509-parser.workspace = true
 
 [dev-dependencies]
+rustls-webpki.workspace = true
 test-utils.workspace = true
 threshold-execution = { workspace = true, features = ["testing"] }
 tfhe = { workspace = true, features = [
