+++ core/threshold/src/choreography/bgv/requests.rs
@@ -1,7 +1,7 @@
 use clap::ValueEnum;
 use serde::{Deserialize, Serialize};
 
-use crate::bgv::basics::LevelEllCiphertext;
+use threshold_experimental::bgv::basics::LevelEllCiphertext;
 use threshold_types::session_id::SessionId;
 
 #[derive(Clone, Debug, ValueEnum, Serialize, Deserialize, PartialEq, Eq, Hash)]
