+++ core/threshold/src/choreography/mod.rs
@@ -1,3 +1,4 @@
-pub mod choreographer;
-pub mod grpc;
-pub mod requests;
+pub mod client_utils;
+pub mod server_utils;
+pub mod tfhe_rs;
+pub mod bgv;
