+++ core/threshold/src/choreography/tfhe_rs/choreographer.rs
@@ -1,53 +1,41 @@
 //! Choreographer is a GRPC client that communicates with
 //! the kms-core (with the moby binary) parties to do benchmarks.
 //! It is a trusted entity and should not be used with production kms-core.
-use crate::choreography::requests::CrsGenParams;
-use crate::conf::choreo::ChoreoConf;
+use super::requests::CrsGenParams;
 use threshold_execution::endpoints::decryption::{DecryptionMode, RadixOrBoolCiphertext};
 use threshold_execution::tfhe_internals::parameters::DkgParamsAvailable;
 use threshold_execution::tfhe_internals::public_keysets::FhePubKeySet;
 use threshold_execution::zk::ceremony::compute_witness_dim;
 
 use algebra::base_ring::Z64;
-use observability::telemetry::ContextPropagator;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use tfhe::xof_key_set::CompressedXofKeySet;
 use threshold_execution::zk::ceremony::InternalPublicParameter;
-use threshold_networking::constants::{MAX_EN_DECODE_MESSAGE_SIZE, NETWORK_TIMEOUT_LONG};
-use threshold_types::party::Identity;
 use threshold_types::role::Role;
 use threshold_types::session_id::SessionId;
 use tokio::{task::JoinSet, time::Duration};
-use tonic::service::interceptor::InterceptedService;
-use tonic::transport::{Channel, Uri};
 use tracing::{Instrument, instrument};
 
 use super::grpc::SupportedRing;
 use super::requests::{
-    PreprocDecryptParams, PreprocKeyGenParams, PrssInitParams, ReshareParams, SessionType, Status,
+    PreprocDecryptParams, PreprocKeyGenParams, PrssInitParams, ReshareParams, SessionType,
     TfheType, ThresholdDecryptParams, ThresholdKeyGenParams, ThresholdKeyGenResultParams,
     ThroughtputParams,
 };
 use threshold_networking::choreography_gen::{
     CrsGenRequest, CrsGenResultRequest, PreprocDecryptRequest, PreprocKeyGenRequest,
-    PrssInitRequest, ReshareRequest, StatusCheckRequest, ThresholdDecryptRequest,
-    ThresholdDecryptResultRequest, ThresholdKeyGenRequest, ThresholdKeyGenResultRequest,
-    choreography_client::ChoreographyClient,
+    PrssInitRequest, ReshareRequest, ThresholdDecryptRequest, ThresholdDecryptResultRequest,
+    ThresholdKeyGenRequest, ThresholdKeyGenResultRequest,
 };
 
-pub struct ChoreoRuntime {
-    pub role_assignments: HashMap<Role, Identity>,
-    pub channels: HashMap<Role, Channel>,
-}
+pub use crate::choreography::client_utils::ChoreoRuntime;
 
 #[derive(Debug)]
 pub struct GrpcOutputs {
     pub outputs: HashMap<String, Z64>,
     pub elapsed_times: Option<HashMap<Role, Vec<Duration>>>,
 }
-
-pub type NetworkTopology = HashMap<Role, Uri>;
 
 #[allow(clippy::large_enum_variant)]
 #[derive(Serialize, Deserialize, Clone)]
@@ -57,53 +45,6 @@ pub enum KeySetMaybeCompressed {
 }
 
 impl ChoreoRuntime {
-    pub fn new_from_conf(conf: &ChoreoConf) -> Result<ChoreoRuntime, Box<dyn std::error::Error>> {
-        let topology = &conf.threshold_topology;
-
-        let role_assignments: HashMap<Role, Identity> = topology.into();
-
-        // we need to set the protocol in URI correctly
-        // depending on whether the certificates are present
-        let host_channels = topology.choreo_physical_topology_into_network_topology()?;
-
-        ChoreoRuntime::new_with_net_topology(role_assignments, host_channels)
-    }
-
-    fn new_with_net_topology(
-        role_assignments: HashMap<Role, Identity>,
-        network_topology: NetworkTopology,
-    ) -> Result<ChoreoRuntime, Box<dyn std::error::Error>> {
-        let channels = network_topology
-            .iter()
-            .map(|(role, host)| {
-                let endpoint: &Uri = host;
-                println!("connecting to endpoint: {:?}", endpoint);
-                // Use the TLS_NODELAY mode to ensure everything gets sent immediately by disabling Nagle's algorithm.
-                // Note that this decreases latency but increases network bandwidth usage. If bandwidth is a concern,
-                // then this should be changed
-                let channel = Channel::builder(endpoint.clone())
-                    .timeout(NETWORK_TIMEOUT_LONG)
-                    .tcp_nodelay(true)
-                    .connect_lazy();
-                Ok((*role, channel))
-            })
-            .collect::<Result<_, Box<dyn std::error::Error>>>()?;
-
-        Ok(ChoreoRuntime {
-            role_assignments,
-            channels,
-        })
-    }
-
-    pub fn new_client(
-        &self,
-        channel: Channel,
-    ) -> ChoreographyClient<InterceptedService<Channel, ContextPropagator>> {
-        ChoreographyClient::with_interceptor(channel, ContextPropagator)
-            .max_decoding_message_size(*MAX_EN_DECODE_MESSAGE_SIZE)
-            .max_encoding_message_size(*MAX_EN_DECODE_MESSAGE_SIZE)
-    }
-
     #[instrument(name = "PRSS-INIT Request", skip(self, session_id), fields(sid = ?session_id))]
     pub async fn inititate_prss_init(
         &self,
@@ -629,50 +570,5 @@ impl ChoreoRuntime {
         }
 
         Ok(*ref_response)
-    }
-
-    pub async fn initiate_status_check(
-        &self,
-        session_id: SessionId,
-        retry: bool,
-        interval: Duration,
-        malicious_roles: Vec<Role>,
-    ) -> anyhow::Result<Vec<(Role, Status)>> {
-        let mut join_set = JoinSet::new();
-        let serialized_sid = bc2wrap::serialize(&session_id)?;
-        let request = StatusCheckRequest {
-            request_id: serialized_sid,
-        };
-        let mut result = Vec::new();
-        loop {
-            self.channels.iter().for_each(|(role, channel)| {
-                let mut client = self.new_client(channel.clone());
-
-                let request = request.clone();
-                let role = *role;
-                join_set.spawn(async move { (role, client.status_check(request).await) });
-            });
-
-            while let Some(response) = join_set.join_next().await {
-                let (role, response) = response?;
-                let status: Status = bc2wrap::deserialize_safe(&response?.into_inner().status)?;
-                result.push((role, status));
-            }
-
-            if !retry
-                || result.iter().all(|(role, status)| {
-                    *status != Status::Ongoing || malicious_roles.contains(role)
-                })
-            {
-                return Ok(result);
-            } else {
-                println!("Status Check for Session ID {session_id} -- Still have running parties");
-                result.sort_by_key(|(role, _)| role.one_based());
-                for (role, status) in result.drain(..) {
-                    println!("Role {role}, Status {status:?}");
-                }
-                tokio::time::sleep(interval).await;
-            }
-        }
     }
 }
