+++ core/threshold/src/conf/choreo.rs
@@ -3,12 +3,13 @@
 //!
 use std::collections::HashMap;
 
-use crate::choreography::choreographer::NetworkTopology;
 use observability::conf::TelemetryConfig;
 use serde::{Deserialize, Serialize};
 use threshold_types::party::Identity;
 use threshold_types::role::Role;
 use tonic::transport::Uri;
+
+pub type NetworkTopology = HashMap<Role, Uri>;
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct ChoreoParty {
