+++ core/threshold/src/grpc/mod.rs
@@ -1,2 +1,1 @@
-#[cfg(feature = "choreographer")]
 pub mod server;
