+++ core/threshold/src/lib.rs
@@ -1,4 +1,4 @@
-#[cfg(all(feature = "choreographer", feature = "non-wasm"))]
+#[cfg(feature = "non-wasm")]
 pub mod choreography;
 #[cfg(feature = "non-wasm")]
 pub use tokio;
@@ -8,14 +8,6 @@ pub mod allocator;
 pub mod conf;
 #[cfg(feature = "non-wasm")]
 pub mod grpc;
-
-#[cfg(feature = "non-wasm")]
-pub mod tls_certs;
-
 pub mod utils;
-
-#[cfg(all(feature = "choreographer", feature = "non-wasm"))]
-pub mod malicious_moby;
-
 #[cfg(feature = "non-wasm")]
 pub mod zk_utils;
