+++ core/threshold/test_scripts/bgv_test_script_fake_dkg.sh
