+++ core/threshold/test_scripts/bgv_test_script_real_dkg.sh
